Refactor api into base
