+++ sdks/python/apache_beam/ml/inference/pytorch_inference.py
@@ -27,7 +27,7 @@ from typing import Union
 
 import torch
 from apache_beam.io.filesystems import FileSystems
-from apache_beam.ml.inference.api import PredictionResult
+from apache_beam.ml.inference.base import PredictionResult
 from apache_beam.ml.inference.base import InferenceRunner
 from apache_beam.ml.inference.base import ModelLoader
