+++ sdks/python/apache_beam/ml/inference/sklearn_inference.py
@@ -25,7 +25,7 @@ import numpy
 from sklearn.base import BaseEstimator
 
 from apache_beam.io.filesystems import FileSystems
-from apache_beam.ml.inference.api import PredictionResult
+from apache_beam.ml.inference.base import PredictionResult
 from apache_beam.ml.inference.base import InferenceRunner
 from apache_beam.ml.inference.base import ModelLoader
